Consolidate HTTP telemetry tests
